Module Framac_ival

module Abstract_interp : sig ... end

Functors for generic lattices implementations. @plugin developer guide

module Bottom : sig ... end

Types, monads and utilitary functions for lattices in which the bottom is managed separately from other values.

module Fc_float : sig ... end

Implementation of floating-point values of different precision, using the standard ocaml floating-point numbers in double precision. Long_Double and Real are inexact.

module Float_interval : sig ... end

module Float_interval_sig : sig ... end

Signature for the floating-point interval semantics.

module Float_sig : sig ... end

Interface of floating-point numbers of different precisions.

module Floating_point : sig ... end

Floating-point operations.

module Fval : sig ... end

Floating-point intervals, used to construct arithmetic lattices. The interfaces of this module may change between Frama-C versions. Contact us if you need stable APIs.

module Integer : sig ... end

Extension of Big_int compatible with Zarith.

module Ival : sig ... end

Arithmetic lattices. The interfaces of this module may change between Frama-C versions. Contact us if you need stable APIs.

module Ival_noinf : sig ... end

Arithmetic lattices. The interfaces of this module may change between Frama-C versions. Contact us if you need stable APIs.

module Pretty_utils : sig ... end