Module Abstract_interp.Comp

Signatures for comparison operators ==, !=, <, >, <=, >=.

type t =

| Lt

| Gt

| Le

| Ge

| Eq

| Ne
(*
comparison operators
*)

type result = truth =

| True

| False

| Unknown
(*
result of a comparison
*)

val pretty_comp : Format.formatter -> t -> unit

val inv : t -> t

Inverse relation: a op b <==> ! (a (inv op) b).

val sym : t -> t

Opposite relation: a op b <==> b (sym op) a.

val inv_result : result -> result

Given a result r for an operation op, inv_result r is the result that would have been obtained for inv op.