Module Framac_ival.Pretty_utils

type sformat = (unit, Format.formatter, unit) format

type 'a formatter = Format.formatter -> 'a -> unit

val pp_iter : ?pre:sformat -> ?sep:sformat -> ?suf:sformat -> (('a -> unit) -> 'b -> unit) -> 'a formatter -> 'b formatter

pretty prints any structure using an iterator on it. The argument pre (resp. suf) is output before (resp. after) the iterator is started (resp. has ended). The optional argument sep is output between two calls to the 'a formatter. Default: open a box for pre, close a box for suf, nothing for sep.